# perf: #62: Defer heavy imports for info commands