## Summary
Optimizes startup performance for info commands (--help, --version, --list-sessions) by deferring heavy imports until after argument parsing.

Fixes #62 

## Changes
- Moved imports of `create_app`, `init_logging`, and `load_dotenv` to after argument parsing in `main.py`
- Added early exit paths for info commands before full app initialization
- Optimized `--list-sessions` to use minimal imports (though still limited by ADK dependencies)

## Performance Results
- `streetrace --help`: **~0.3s** (was 2.5s) - **88% improvement** ✅
- `streetrace --version`: **~0.3s** (was 2.5s) - **88% improvement** ✅
- `streetrace --list-sessions`: ~2.8s (limited by Google ADK imports)

## Testing
- [x] All tests pass (`make test`)
- [x] Linting passes (`make lint`)
- [x] Performance verified with `time` command
- [x] No regression in normal application usage

## Notes
- The main remaining bottleneck is the Google ADK framework import (~1.7s)
- Further optimization would require lazy loading of the ADK framework or architectural changes
- Added `profile_output.prof` to `.gitignore` for future profiling work