@wassertim 2025-07-27
Closing this PR as our investigation showed it doesn't actually improve startup performance. The deferred imports don't affect the real 2.5s bottleneck which occurs earlier in the import chain. Will focus on properly investigating the root cause in issue #31.